backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
google-generativeai
python-dotenv
slowapi
sse-starlette
